requirements.txt
flask>=2.3.0
pyyaml>=6.0
pandas>=1.5.0